@Julian Conner Added Paths for game, some Ghost movement classes
I also touched up the Position2D class and made the increment() & decrement()  methods for x and y
more efficient in the Entity class. The Ghost movement algorithms can be set and swapped out anytime
before or during the program by setting the Ghost's GhostMovementStrategy instance variable to an instance of its subclass that
has the movement algrorithm we want. By default it will use an object that doesn't do anything but
be a placeholder so the instance variable is never null, so it won't crash the program if someone forgot
to set it. It also saves us time because we don't have to keep checking for null each time we want to
use the GhostMovementStrategy object.

Future : Will make a fully connected set of Path objects for a real game board design.
             Will make different shortest path and maze solver algorithms for the Ghosts to use.