Made Ghost movements work with the Hit Detection
Fixed Error : The Ghost movements didn't stop at the wall.
Reason : The increment() decrement() methods for x and y individually update the Ghost's previous
               position and the Ghost movement algorithm called one of those methods for both x and
               y, which changed the Ghost position twice. This saved the previous position to the
               first change in position instead of before both changes in position.

              This was an issue because the hit detection algorithm set the Ghost to the saved previous                  point after it hit a wall , which allowed one increment or decrement towards Pacman
               each time because the previous position was saved as the spot after the first x or y
               movement, instead as before the both x and y were moved.

 Solution: I made the GhostMovementStrategy class save the Ghost position before the algorithm
                was applied and set it as the Ghost's previous position after the algorithm was finished.

Details About Changes :

I made the GhostMovementStrategy abstract class have a move() method that automatically
saves the last position before the movement algorithm, performs the algorithm method, then
updates the Ghost's previous position with the saved position after the movement algorithm.
I changed the abstract method to a movementAlgorithm() method so all the movement
subclasses only have to implement the algorithm. I added the ability to set whether or not
the Ghost is moving or staying still. If the Ghost is set to stay still then the move() function
doesn't perform the algorithm.

I modified some of the code in the SystemManager and Helper class to make it more readable.